cmpt354_mp/backend/requirements.txt
alembic==1.15.2
APScheduler==3.11.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
Flask==3.0.0
Flask-APScheduler==1.13.1
Flask-Cors==4.0.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
pycparser==2.22
python-dateutil==2.9.0.post0
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.0
tzlocal==5.3.1
Werkzeug==3.1.3
zipp==3.21.0
